Used Tie::IxHash to get ordered plugins.
In our use cases the order of the plugins can be important, so we would
like to be able to use the order the plugins are added to the object to
determine the order they are returned from the plugin_list method. Using
Tie::IxHash provides this functionality, but would force all calls to
plugin_list to be ordered. For those people who don't care about plugin
order, this shouldn't mater anyway.